+++ src/main/java/main/FastEnemy.java
@@ -6,7 +6,7 @@ public class FastEnemy extends GameObject {
 
     private Handler handler;
 
-    public FastEnemy(int x, int y, ID id, Handler handler) {
+    public FastEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
         velX = 2;
         velY = 9;
@@ -32,11 +32,11 @@ public class FastEnemy extends GameObject {
     @Override
     public void render(Graphics g) {
         g.setColor(Color.CYAN);
-        g.fillRect(x,y,16,16);
+        g.fillRect((int)x,(int)y,16,16);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16,16);
+        return new Rectangle((int)x, (int)y, 16,16);
     }
 }
